Remove unneeded columns from protocol table in esf and add status column
The protocol table contained columns for login-name, name, studypath, matrikel
of the student, but this data was not actually depending on the protocol, the
columns simply showed the global values of the student.
In contrast, the current status of the exam (registered, earmarked) was not
displayed, so added a cloumn for this.
Additionally the code was restructured and modernized.