Create JPA entity ArchivedProtocol and manager to archive old exam protocols
This new entity will be display in the user's esf and can be used to display
exam results wihtout the need to have a relation to any exam or protocol entity.
The mapping is done by the user's student number.
Currently there is no way to import/export or manually add/edit these protocols.