Update the esf student and edit controller to show archived protocols
To compensate for the additional usage of vertical space, we trim down the header
showing the personal information of the student by showing two colums and floating
the edit button and we reduce the number of rows in the tables.
Additionally fix some problems in the controllers and clean up code.