Rework exam details view: Slim layout, remaining time as text and status colors.
The old tabular view was replaced with an only two line table that compresses
the information and tries to make it visually comprehensible.
The remaining time for registrations or unregistrations is now additionally shown
as text (eg 3 days remaining) and the two phases are highlighted green when they
are active and red when they are over.
Boolean attributes are now shown as (disabled) checkboxes.