Change login button icons to fit our use case
We now use the university icon for university members and the graduation cap for
students. The text of the olat login was also changed to "Mitarbeiter".
Note that the used class names are not semantic anymore, but otherwise we would
have to define our own theme or modify the existing one which would include
recompilation using sass.

Conflicts:
	src/main/java/org/olat/login/_i18n/LocalStrings_de.properties