Display esf in a separate tab
Previously the esf was displayed in the portal. In OO 10 the portal is disabled
by default and the menu entries have been moved to the user dropdown menu, where
the esf would be difficult to find.
So instead we create a separate site for the esf that is disabled as a permanent
tab in the titlebar and can be customized using the admin menu.