Fix min/max score in course and structure nodes
These used to throw RuntimeExceptions instead of just returning null. Throwing
exceptions is only allowed if hasScoreConfigured returns false, which is not
the case, because both actually provide a score.
But min oder max scores do not make sense, because generally a lesser score is
better in autotool (eg. smaller fsm).