Workaround for failed signature check
Sometimes the signature check fails because the redstone XML-RPC trims leading
and trailing whitespace in responses, which results in different signatures.
To workaround this, we disable trimming completely.

Because the redstone library does not support turning off trimming, we switch
to an accordingly modified fork of the latest version.
See branch autolat in https://github.com/klemens/redstone for more details.