Improve performance of structure node
This commit adds an identity->student cache to the highscore and pointoverview
tables that prevents a raw db lookup on every getValueAt call.
This greatly improves performance, especially for courses with a large number
of enrolled users, because the mentioned tables get sorted indirectly by users.

The two tables could be improved even more to the extent that the whole table
gets precalculated in memory.