Fix context help after the local help system got removed
OpenOLAT 10.4 removed the legacy help system that allowed small localised
context help pages that were rendered automatically and replaced it with a
system that links directly to a externally hosted confluence wiki.

Because of the lack of a local help system, this commit merges all help
messages into one html doc that uses http fragments for linking to the
relevant sections. Additionally, the default help system of OO is replaced
with one that links to the mentioned html doc and falls back to the default
system for all other links.

This help system must be enabled in olat.local.properties using the property
help.plugin=autolatHelp. The system can also be chained with yet another help
system using the autolat.help.fallback property.