Fix compilation error in the structure node
The autolat structure node, which was copied from the standard structure node
when it was originally created, used a old api that was removed in the latest
OO update.

This commit is a quick and dirty fix for the immediate problem, but as it
turns out, the structure node itself is really broken and should be updated
(or replaced with the standard structure node!).