Fix exception due to stale data when listing exam authors
This was caused by referencing the user request object in an anonymous class
to get the identity, which causes it to be copied into the class instance.
Getting the identity on further requests then fails because the object is no
longer valid. We avoid this by direcly referencing the identity object.