Always use random seed for the get_task_instance RPC call
Previously, we were using the student's institutional number or a string
derived from it as a seed when requesting new living task instances. This
lead to problems where the same task instance was returned for different
task configurations, because the autotool server is caching the task
instances by seed.

Now, we always generate a 20 character, random, alphanumeric seed (~119 b).
We don't have to save this seed, as we are already saving the returned task
instance in the database.